chore: move new serializers to v2
