+++ course_discovery/apps/course_metadata/search_indexes/serializers/aggregation.py
@@ -5,19 +5,16 @@ from rest_framework.serializers import ListSerializer
 from course_discovery.apps.api.serializers import COMMON_IGNORED_FIELDS
 from course_discovery.apps.course_metadata.search_indexes import documents
 from course_discovery.apps.edx_elasticsearch_dsl_extensions.serializers import (
-    BaseDjangoESDSLFacetSerializer, DummyDocument, MultiDocumentSerializerMixin
+    BaseDjangoESDSLFacetSerializer,
+    DummyDocument,
+    MultiDocumentSerializerMixin,
 )
 
-from .course import CourseSearchDocumentSerializer, CourseSearchDocumentSerializerV2, CourseSearchModelSerializer
-from .course_run import (
-    CourseRunSearchDocumentSerializer, CourseRunSearchDocumentSerializerV2, CourseRunSearchModelSerializer
-)
-from .learner_pathway import (
-    LearnerPathwaySearchDocumentSerializer, LearnerPathwaySearchDocumentSerializerV2,
-    LearnerPathwaySearchModelSerializer
-)
-from .person import PersonSearchDocumentSerializer, PersonSearchDocumentSerializerV2
-from .program import ProgramSearchDocumentSerializer, ProgramSearchDocumentSerializerV2, ProgramSearchModelSerializer
+from .course import CourseSearchDocumentSerializer, CourseSearchModelSerializer
+from .course_run import CourseRunSearchDocumentSerializer, CourseRunSearchModelSerializer
+from .learner_pathway import LearnerPathwaySearchDocumentSerializer, LearnerPathwaySearchModelSerializer
+from .person import PersonSearchDocumentSerializer
+from .program import ProgramSearchDocumentSerializer, ProgramSearchModelSerializer
 
 
 class AggregateSearchModelSerializer(MultiDocumentSerializerMixin, DocumentSerializer):
@@ -115,27 +112,6 @@ class AggregateSearchListSerializer(MultiDocumentSerializerMixin, ListSerializer
         }
 
 
-class AggregateSearchListSerializerV2(AggregateSearchListSerializer):
-    """
-    Extended version of the AggregateSearchListSerializer with updated serializers that support search_after pagination.
-
-    This subclass allows for the use of newer serializer versions for the same document types,
-    which include additional search index fields specifically for version 2.
-    """
-
-    class Meta(AggregateSearchListSerializer.Meta):
-        """
-        Meta options.
-        """
-        serializers = {
-            documents.CourseRunDocument: CourseRunSearchDocumentSerializerV2,
-            documents.CourseDocument: CourseSearchDocumentSerializerV2,
-            documents.ProgramDocument: ProgramSearchDocumentSerializerV2,
-            documents.LearnerPathwayDocument: LearnerPathwaySearchDocumentSerializerV2,
-            documents.PersonDocument: PersonSearchDocumentSerializerV2,
-        }
-
-
 class AggregateSearchSerializer(DocumentSerializer):
     """
     Serializer for aggregated elasticsearch documents.
@@ -150,14 +126,3 @@ class AggregateSearchSerializer(DocumentSerializer):
         ignore_fields = COMMON_IGNORED_FIELDS
 
 
-class AggregateSearchSerializerV2(AggregateSearchSerializer):
-    """
-    Serializer for aggregated elasticsearch documents.
-    """
-
-    class Meta(AggregateSearchSerializer.Meta):
-        """
-        Meta options.
-        """
-        list_serializer_class = AggregateSearchListSerializerV2
-        document = DummyDocument
