+++ course_discovery/apps/course_metadata/search_indexes/serializers/course.py
@@ -14,10 +14,10 @@ from course_discovery.apps.api.serializers import ContentTypeSerializer, CourseW
 from course_discovery.apps.course_metadata.utils import get_course_run_estimated_hours, get_product_skill_names
 from course_discovery.apps.edx_elasticsearch_dsl_extensions.serializers import BaseDjangoESDSLFacetSerializer
 
-from ..constants import BASE_SEARCH_INDEX_FIELDS, COMMON_IGNORED_FIELDS, SEARCH_INDEX_ADDITIONAL_FIELDS_V2
+from ..constants import BASE_SEARCH_INDEX_FIELDS, COMMON_IGNORED_FIELDS
 from ..documents import CourseDocument
 from .common import (
-    DateTimeSerializerMixin, DocumentDSLSerializerMixin, ModelObjectDocumentSerializerMixin, SortFieldMixin
+    DateTimeSerializerMixin, DocumentDSLSerializerMixin, ModelObjectDocumentSerializerMixin
 )
 
 __all__ = ('CourseSearchDocumentSerializer',)
@@ -218,21 +218,6 @@ class CourseSearchDocumentSerializer(ModelObjectDocumentSerializerMixin, DateTim
             'external_course_marketing_type',
             'product_source',
         )
-
-
-class CourseSearchDocumentSerializerV2(SortFieldMixin, CourseSearchDocumentSerializer):
-    """
-    Serializer for Course documents, extending the base `CourseSearchDocumentSerializer`
-    to include additional fields for enhanced search functionality, as well as a `sort` field
-    to provide sorting information from the Elasticsearch response.
-
-    This serializer expands the `fields` attribute in the `Meta` class to include additional
-    fields specified in `SEARCH_INDEX_ADDITIONAL_FIELDS_V2`.
-    """
-
-    class Meta(CourseSearchDocumentSerializer.Meta):
-        document = CourseDocument
-        fields = CourseSearchDocumentSerializer.Meta.fields + SEARCH_INDEX_ADDITIONAL_FIELDS_V2
 
 
 # pylint: disable=abstract-method
