+++ course_discovery/apps/course_metadata/search_indexes/serializers/learner_pathway.py
@@ -8,10 +8,10 @@ from course_discovery.apps.api.fields import StdImageSerializerField
 from course_discovery.apps.api.serializers import ContentTypeSerializer
 from course_discovery.apps.learner_pathway.api.serializers import LearnerPathwaySerializer, LearnerPathwayStepSerializer
 
-from ..constants import BASE_SEARCH_INDEX_FIELDS, COMMON_IGNORED_FIELDS, SEARCH_INDEX_ADDITIONAL_FIELDS_V2
+from ..constants import BASE_SEARCH_INDEX_FIELDS, COMMON_IGNORED_FIELDS
 from ..documents import LearnerPathwayDocument
 from .common import (
-    DateTimeSerializerMixin, DocumentDSLSerializerMixin, ModelObjectDocumentSerializerMixin, SortFieldMixin
+    DateTimeSerializerMixin, DocumentDSLSerializerMixin, ModelObjectDocumentSerializerMixin
 )
 
 __all__ = ('LearnerPathwaySearchDocumentSerializer',)
@@ -79,21 +79,6 @@ class LearnerPathwaySearchDocumentSerializer(
             setattr(instance, 'object', _object)  # pylint: disable=literal-used-as-attribute
             prepared_instance = instance
         return super().to_representation(prepared_instance)
-
-
-class LearnerPathwaySearchDocumentSerializerV2(SortFieldMixin, LearnerPathwaySearchDocumentSerializer):
-    """
-    Serializer for Learner Pathway documents, extending the base `LearnerPathwaySearchDocumentSerializer`
-    to include additional fields for enhanced search functionality, as well as a `sort` field
-    to provide sorting information from the Elasticsearch response.
-
-    This serializer expands the `fields` attribute in the `Meta` class to include additional
-    fields specified in `SEARCH_INDEX_ADDITIONAL_FIELDS_V2`.
-    """
-
-    class Meta(LearnerPathwaySearchDocumentSerializer.Meta):
-        document = LearnerPathwayDocument
-        fields = LearnerPathwaySearchDocumentSerializer.Meta.fields + SEARCH_INDEX_ADDITIONAL_FIELDS_V2
 
 
 class LearnerPathwaySearchModelSerializer(DocumentDSLSerializerMixin, ContentTypeSerializer, LearnerPathwaySerializer):
